+++ packages/core/lib/match.js
@@ -9,7 +9,7 @@ module.exports = match
  * like `group()`, but adds the file matching information to the context.
  *
  * @param {string|RegExp|Function|Array} test   A glob like `*.css` or `{*.js, *.jsx}` or something else to use as `loader.test`.
- * @param {object} [options]                    Optional advanced matching options.
+ * @param {object} [options]                    Rule options. See https://webpack.js.org/configuration/module/
  * @param {Function[]} configSetters            Array of functions as returned by webpack blocks.
  * @return {Function}
  */
@@ -20,6 +20,10 @@ function match (test, options, configSetters) {
   }
 
   assertConfigSetters(configSetters)
+
+  if (options.test) {
+    console.warn(`match(): Setting 'test' in options is not supported and will be overriden with a 'test' argument.`)
+  }
 
   options.test = createFileTypeMatcher(test)
 
@@ -36,14 +40,14 @@ const regexify = glob => globToRegex(glob, { extended: true })
 function createFileTypeMatcher (test) {
   if (typeof test === 'string') {
     return regexify(test)
-  } else if (Array.isArray(test) && test.every(item => typeof item === 'string')) {
-    return test.map(item => regexify(item))
-  } else {
-    return test
   }
+  if (Array.isArray(test) && test.every(item => typeof item === 'string')) {
+    return test.map(item => regexify(item))
+  }
+  return test
 }
 
-function deriveContextWithMatch (context, match) {
+function deriveContextWithMatch (context, options) {
   // Use an ES Proxy, so the returned context will include `match` and mutations
   // to it will also be applied to the original context. This is important,
   // since we have now got a global context object and local ones.
@@ -52,7 +56,7 @@ function deriveContextWithMatch (context, match) {
       return propName === 'match' ? true : (propName in target)
     },
     get (target, propName) {
-      return propName === 'match' ? match : target[propName]
+      return propName === 'match' ? options : target[propName]
     }
   })
 }